requirements.txt
streamlit
PyPDF2
python-docx
google-generativeai
python-dotenv
fpdf2
pandas